Merging r196397: ------------------------------------------------------------------------ r196397 | gclayton | 2013-12-04 10:53:50 -0800 (Wed, 04 Dec 2013) | 7 lines
Added a new directory type for the "bool Host::GetLLDBPath (PathType path_type, FileSpec &file_spec)" function: ePathTypeLLDBTempSystemDir

This will get the temporary directory on the current system.

Removed a call to tmpnam() and replaced it with a call to mktemp() using a template that will be in the temp directory.


------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project@196598 91177308-0d34-0410-b5e6-96231b3b80d8