Update Clang's CFGBlock interface to conform to the strange part of LLVM's Value interface which is used in LLVM's DominatorTree analysis and which changed in LLVM r198836.
The DominatorTree analysis is actually a generic graph analysis and
should be moved to LLVM's support library to clarify that Clang and
others are using it with arbitrary graphs. Further, it seems likely that
it should be using something other than printAsOperand, but this is
a simpler build fix. I'll clean this up later.

git-svn-id: https://llvm.org/svn/llvm-project@198840 91177308-0d34-0410-b5e6-96231b3b80d8