Sane default for --commit in lnt submit
Lnt submit is used for sending results to the server, so having the default behavior be not submitting is confusing to users.  This patch changes the default to commit=1, so results are submitted, and adds a warning if submit is called with commit=0 so that you know your results will not be stored on the server.

git-svn-id: https://llvm.org/svn/llvm-project@199271 91177308-0d34-0410-b5e6-96231b3b80d8