AAPCS: Do not split structs after CPRC allocated on stack
According to the AAPCS, we can split structs between GPRs and the stack,
except for when an argument has already been allocated on the stack. This
can occur when a large number of floating-point arguments fill up the VFP
registers, and are alllocated on the stack before the general-purpose argument
registers are full.



git-svn-id: https://llvm.org/svn/llvm-project@201137 91177308-0d34-0410-b5e6-96231b3b80d8