Keep the link register for uwtable.
The function with uwtable attribute might be visited by the
stack unwinder, thus the link register should be considered
as clobbered after the execution of the branch and link
instruction (i.e. the definition of the machine instruction
can't be ignored) even when the callee function are marked
with noreturn.


git-svn-id: https://llvm.org/svn/llvm-project@202165 91177308-0d34-0410-b5e6-96231b3b80d8