[PM] As Dave noticed in review, I had erroneously copied the move constructors from the classes which only have a single reference member to many other places. This resulted in them copying their single member instead of moving. =/ Fix this.
There's really not a useful test to add sadly because these move
constructors are only called when something deep inside some standard
library implementation *needs* to move them. Many of the types aren't
even user-impacting types. Or, the objects are copyable anyways and so
the result was merely a performance problem rather than a correctness
problem.

Anyways, thanks for the review. And this is a great example of why
I wish I colud have the compiler write these for me.

git-svn-id: https://llvm.org/svn/llvm-project@203431 91177308-0d34-0410-b5e6-96231b3b80d8