lex: improve include handling on Linux for Windows
Normalise the path separator character on non-windows platforms.  Although this
would work on Windows as well (most newer versions of Windows support either '/'
or '\' as a path separator character), it could potentially cause problems with
full UNC paths.  This change enables the use of the Windows SDK on Linux which
will not accept '\' as a path separator.

git-svn-id: https://llvm.org/svn/llvm-project@203614 91177308-0d34-0410-b5e6-96231b3b80d8