move WITH_POLLY option before add_subdirectory(tools)
the first run of the polly buildbot failed, and then it started passing.
This is due to the fact that the buildbot re-builds in an existing directory,
and the first run does not have WITH_POLLY set when it enters tools/.
Thus, cmake ignores the tools/polly dir in the first run, and then because
it reuses the CMakeCache.txt of the previous run, it has the WITH_POLLY set
by the previous run, and so it passes the second time.

git-svn-id: https://llvm.org/svn/llvm-project@203615 91177308-0d34-0410-b5e6-96231b3b80d8